✨ refactor: update imports for EditorGroupTabInfo
Refactor the import statements across multiple files to use thecorrect package for `EditorGroupTabInfo`. This change improves 
code organization and ensures consistency in the usage of 
the `EditorGroupTabInfo` class throughout the project.
