+++ src/main/java/krasa/editorGroups/tabs2/EditorGroupsTabsListener.kt
@@ -1,5 +1,7 @@
 package krasa.editorGroups.tabs2
 
+import krasa.editorGroups.tabs2.label.EditorGroupTabInfo
+
 interface EditorGroupsTabsListener {
   fun selectionChanged(oldSelection: EditorGroupTabInfo?, newSelection: EditorGroupTabInfo?): Unit = Unit
   fun beforeSelectionChanged(oldSelection: EditorGroupTabInfo?, newSelection: EditorGroupTabInfo?): Unit = Unit
