+++ src/main/java/krasa/editorGroups/tabs2/impl/KrLayoutPassInfo.java
@@ -1,7 +1,7 @@
 // Copyright 2000-2019 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 package krasa.editorGroups.tabs2.impl;
 
-import krasa.editorGroups.tabs2.EditorGroupTabInfo;
+import krasa.editorGroups.tabs2.label.EditorGroupTabInfo;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
