+++ src/main/java/krasa/editorGroups/tabs2/impl/KrTabLabel.java
@@ -20,10 +20,10 @@ import com.intellij.util.ObjectUtils;
 import com.intellij.util.ui.*;
 import com.intellij.util.ui.accessibility.ScreenReader;
 import krasa.editorGroups.settings.EditorGroupsSettings;
-import krasa.editorGroups.tabs2.EditorGroupTabInfo;
 import krasa.editorGroups.tabs2.EditorGroupsTabsEx;
 import krasa.editorGroups.tabs2.impl.painter.KrTabPainterAdapter;
 import krasa.editorGroups.tabs2.impl.themes.EditorGroupTabTheme;
+import krasa.editorGroups.tabs2.label.EditorGroupTabInfo;
 import krasa.editorGroups.tabs2.label.TabUiDecorator;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
