+++ src/main/java/krasa/editorGroups/tabs2/impl/multiRow/KrSimpleTabsRow.kt
@@ -1,6 +1,6 @@
 package krasa.editorGroups.tabs2.impl.multiRow
 
-import krasa.editorGroups.tabs2.EditorGroupTabInfo
+import krasa.editorGroups.tabs2.label.EditorGroupTabInfo
 
 class KrSimpleTabsRow(
   infos: List<EditorGroupTabInfo>,
