+++ src/main/java/krasa/editorGroups/tabs2/impl/multiRow/KrTabsRow.kt
@@ -1,6 +1,6 @@
 package krasa.editorGroups.tabs2.impl.multiRow
 
-import krasa.editorGroups.tabs2.EditorGroupTabInfo
+import krasa.editorGroups.tabs2.label.EditorGroupTabInfo
 import java.awt.Rectangle
 
 abstract class KrTabsRow(val infos: List<EditorGroupTabInfo>, val withTitle: Boolean, val withEntryPointToolbar: Boolean) {
