+++ src/main/java/krasa/editorGroups/tabs2/impl/singleRow/KrCompressibleSingleRowLayout.java
@@ -2,10 +2,10 @@
 package krasa.editorGroups.tabs2.impl.singleRow;
 
 import com.intellij.ui.tabs.impl.JBTabsImpl;
-import krasa.editorGroups.tabs2.EditorGroupTabInfo;
 import krasa.editorGroups.tabs2.EditorGroupsTabsPosition;
 import krasa.editorGroups.tabs2.impl.KrTabLabel;
 import krasa.editorGroups.tabs2.impl.KrTabsImpl;
+import krasa.editorGroups.tabs2.label.EditorGroupTabInfo;
 
 import java.util.Iterator;
