+++ src/main/java/krasa/editorGroups/tabs2/label/EditorGroupTabInfo.kt
@@ -1,26 +1,9 @@
-/*
- * Copyright 2000-2015 JetBrains s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package krasa.editorGroups.tabs2
+package krasa.editorGroups.tabs2.label
 
 import com.intellij.openapi.editor.markup.TextAttributes
 import com.intellij.openapi.util.NlsContexts
-import com.intellij.openapi.util.NlsContexts.TabTitle
 import com.intellij.ui.SimpleColoredText
 import com.intellij.ui.SimpleTextAttributes
-import com.intellij.ui.SimpleTextAttributes.StyleAttributeConstant
 import krasa.editorGroups.tabs2.impl.KrTabLabel
 import java.awt.Color
 import java.beans.PropertyChangeSupport
@@ -73,7 +56,7 @@ open class EditorGroupTabInfo(var component: JComponent? = null) {
   val coloredText: SimpleColoredText = SimpleColoredText()
 
   /** Text. */
-  var text: @TabTitle String
+  var text: @NlsContexts.TabTitle String
     get() = coloredText.toString()
     set(text) {
       val attributes: MutableList<SimpleTextAttributes?> = coloredText.attributes
@@ -115,7 +98,7 @@ open class EditorGroupTabInfo(var component: JComponent? = null) {
   /** Custom editor attributes. */
   private var editorAttributes: TextAttributes? = null
 
-  @StyleAttributeConstant
+  @SimpleTextAttributes.StyleAttributeConstant
   private var defaultStyle: Int = -1
 
   /** Compute default attributes. */
@@ -159,7 +142,7 @@ open class EditorGroupTabInfo(var component: JComponent? = null) {
     return this
   }
 
-  fun setDefaultStyle(@StyleAttributeConstant style: Int): EditorGroupTabInfo {
+  fun setDefaultStyle(@SimpleTextAttributes.StyleAttributeConstant style: Int): EditorGroupTabInfo {
     defaultStyle = style
     defaultAttributes = null
     update()
